Chrome: Rename windows_utils -> windows_decryption
