+++ monkey/agent_plugins/credentials_collectors/chrome/src/windows_decryption.py
